Task 13 jpa hibernate&spring data (#17)
* Added Jpa database

* Added Jpa database, created tests for SQL database and Jpa database

* Added Jpa database, created tests for SQL database and Jpa database

* Added Jpa database, created tests for SQL database and Jpa database

+++ src/main/java/pl/futurecollars/invoicing/db/memory/InMemoryDatabase.java
@@ -9,18 +9,18 @@ import pl.futurecollars.invoicing.model.Invoice;
 
 public class InMemoryDatabase implements Database {
 
-    private final HashMap<Integer, Invoice> invoices = new HashMap<>();
-    private int index = 1;
+    private final HashMap<Long, Invoice> invoices = new HashMap<>();
+    private long index = 1;
 
     @Override
-    public int save(Invoice invoice) {
+    public long save(Invoice invoice) {
         invoice.setId(index);
         invoices.put(index, invoice);
         return index++;
     }
 
     @Override
-    public Optional<Invoice> getById(int id) {
+    public Optional<Invoice> getById(long id) {
         return Optional.ofNullable(invoices.get(id));
     }
 
@@ -31,14 +31,14 @@ public class InMemoryDatabase implements Database {
     }
 
     @Override
-    public Optional<Invoice> update(int id, Invoice updatedInvoice) {
+    public Optional<Invoice> update(long id, Invoice updatedInvoice) {
         updatedInvoice.setId(id);
         return Optional.ofNullable(invoices.put(id, updatedInvoice));
 
     }
 
     @Override
-    public Optional<Invoice> delete(int id) {
+    public Optional<Invoice> delete(long id) {
         return Optional.ofNullable(invoices.remove(id));
 
     }
